app/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
jsonschema==3.2.0
python-dotenv==0.14.0
requests==2.24.0
typing-extensions==3.7.4.2
super-case-converter-2000
